Start sync analysis!

+++ src/AudioSync/Structures/Onset.cs
@@ -1,3 +1,3 @@
 ﻿namespace AudioSync.Structures;
 
-internal readonly record struct Onset(int Position, float Strength);
+internal readonly record struct Onset(int Position, double Strength);
